Made Sample display of EventDetailView.java a separate file.
